Continue documentation, finalise validations

+++ hce/legacy/atarashii/master/Atarashii/Modules/Profile/Options/Sensitivity.cs
@@ -60,9 +60,13 @@ namespace Atarashii.Modules.Profile.Options
         /// <summary>
         ///     Checks if the inbound sensitivity value falls within a valid range.
         /// </summary>
-        /// <param name="value">Inbound value to check.</param>
-        /// <returns>True on valid value, otherwise false.</returns>
-        private static bool IsValid(int value)
+        /// <param name="value">
+        ///     Inbound value to check.
+        /// </param>
+        /// <returns>
+        ///     True on valid value, otherwise false.
+        /// </returns>
+        private static bool IsValid(ushort value)
         {
             return value > 1 && value < 11;
         }
